requirements.txt
pandas
mysql-connector-python
matplotlib
scikit-learn
streamlit